LRCI-6087 Move all js unit test class group logic
